Refactor user module to implement CQRS pattern and reorganize service imports
