+++ src/user/queries/get-user-profile/get-user-profile.query.ts
@@ -1,0 +1,3 @@
+export class GetUserProfileQuery {
+  constructor(public readonly userId: string) {}
+}
