+++ src/user/user.controller.ts
@@ -6,6 +6,7 @@ import {
   UploadedFile,
   UseInterceptors,
 } from '@nestjs/common';
+import { QueryBus } from '@nestjs/cqrs';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { ApiConsumes, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { BucketService } from 'src/bucket/services/bucket.service';
@@ -14,7 +15,8 @@ import { Auth, CurrentUser } from 'src/common/decorators/auth.decorator';
 import { MulterFile } from 'src/common/interfaces/multer.interface';
 import { UpdateUserDTOWithAvatar } from './dto/update-user.dto';
 import { IUser, IUserDocument } from './interfaces/user.interface';
-import { UserService } from './user.service';
+import { GetUserProfileQuery } from './queries/get-user-profile/get-user-profile.query';
+import { UserService } from './services/user.service';
 
 @Controller('user')
 @ApiTags('User')
@@ -22,13 +24,14 @@ export class UserController {
   constructor(
     private readonly userService: UserService,
     private readonly bucketService: BucketService,
+    private readonly queryBus: QueryBus,
   ) {}
 
   @Get('profile')
   @ApiOperation({ summary: 'Retrieve logged in users profile' })
   @Auth()
   async profile(@CurrentUser() auth: IUserDocument) {
-    return auth;
+    return await this.queryBus.execute(new GetUserProfileQuery(auth.id));
   }
 
   @Patch()
